exception_escaping has renamed to escape
