+++ tests/test_proxy_module.py
@@ -2,14 +2,14 @@ import asyncio
 
 import pytest
 
-import exception_escaping
-from exception_escaping.errors import SetDefaultReturnValueForDecoratorError
+import escape
+from escape.errors import SetDefaultReturnValueForDecoratorError
 
 
 def test_run_simple_function():
     some_value = 'kek'
 
-    @exception_escaping
+    @escape
     def function():
         return some_value
 
@@ -17,7 +17,7 @@ def test_run_simple_function():
 
 
 def test_run_simple_function_with_some_arguments():
-    @exception_escaping
+    @escape
     def function(a, b, c=5):
         return a + b + c
 
@@ -28,7 +28,7 @@ def test_run_simple_function_with_some_arguments():
 
 
 def test_run_function_with_exception():
-    @exception_escaping
+    @escape
     def function(a, b, c=5):
         raise ValueError
 
@@ -38,7 +38,7 @@ def test_run_function_with_exception():
 def test_run_coroutine_function():
     some_value = 'kek'
 
-    @exception_escaping
+    @escape
     async def function():
         return some_value
 
@@ -46,7 +46,7 @@ def test_run_coroutine_function():
 
 
 def test_run_coroutine_function_with_some_arguments():
-    @exception_escaping
+    @escape
     async def function(a, b, c=5):
         return a + b + c
 
@@ -57,7 +57,7 @@ def test_run_coroutine_function_with_some_arguments():
 
 
 def test_run_coroutine_function_with_exception():
-    @exception_escaping
+    @escape
     async def function(a, b, c=5):
         raise ValueError
 
@@ -67,7 +67,7 @@ def test_run_coroutine_function_with_exception():
 def test_run_simple_function_with_empty_brackets():
     some_value = 'kek'
 
-    @exception_escaping()
+    @escape()
     def function():
         return some_value
 
@@ -75,7 +75,7 @@ def test_run_simple_function_with_empty_brackets():
 
 
 def test_run_simple_function_with_some_arguments_with_empty_brackets():
-    @exception_escaping()
+    @escape()
     def function(a, b, c=5):
         return a + b + c
 
@@ -86,7 +86,7 @@ def test_run_simple_function_with_some_arguments_with_empty_brackets():
 
 
 def test_run_function_with_exception_with_empty_brackets():
-    @exception_escaping()
+    @escape()
     def function(a, b, c=5):
         raise ValueError
 
@@ -96,7 +96,7 @@ def test_run_function_with_exception_with_empty_brackets():
 def test_run_coroutine_function_with_empty_brackets():
     some_value = 'kek'
 
-    @exception_escaping()
+    @escape()
     async def function():
         return some_value
 
@@ -104,7 +104,7 @@ def test_run_coroutine_function_with_empty_brackets():
 
 
 def test_run_coroutine_function_with_some_arguments_with_empty_brackets():
-    @exception_escaping()
+    @escape()
     async def function(a, b, c=5):
         return a + b + c
 
@@ -115,7 +115,7 @@ def test_run_coroutine_function_with_some_arguments_with_empty_brackets():
 
 
 def test_run_coroutine_function_with_exception_with_empty_brackets():
-    @exception_escaping()
+    @escape()
     async def function(a, b, c=5):
         raise ValueError
 
@@ -125,7 +125,7 @@ def test_run_coroutine_function_with_exception_with_empty_brackets():
 def test_run_simple_function_with_default_return():
     some_value = 'kek'
 
-    @exception_escaping(default_return='lol')
+    @escape(default_return='lol')
     def function():
         return some_value
 
@@ -133,7 +133,7 @@ def test_run_simple_function_with_default_return():
 
 
 def test_run_simple_function_with_some_arguments_with_default_return():
-    @exception_escaping(default_return='lol')
+    @escape(default_return='lol')
     def function(a, b, c=5):
         return a + b + c
 
@@ -146,7 +146,7 @@ def test_run_simple_function_with_some_arguments_with_default_return():
 def test_run_function_with_exception_with_default_return():
     default_value = 13
 
-    @exception_escaping(default_return=default_value)
+    @escape(default_return=default_value)
     def function(a, b, c=5):
         raise ValueError
 
@@ -156,7 +156,7 @@ def test_run_function_with_exception_with_default_return():
 def test_run_coroutine_function_with_default_return():
     some_value = 'kek'
 
-    @exception_escaping(default_return='lol')
+    @escape(default_return='lol')
     async def function():
         return some_value
 
@@ -164,7 +164,7 @@ def test_run_coroutine_function_with_default_return():
 
 
 def test_run_coroutine_function_with_some_arguments_with_default_return():
-    @exception_escaping(default_return='lol')
+    @escape(default_return='lol')
     async def function(a, b, c=5):
         return a + b + c
 
@@ -177,7 +177,7 @@ def test_run_coroutine_function_with_some_arguments_with_default_return():
 def test_run_coroutine_function_with_exception_with_default_return():
     default_value = 13
 
-    @exception_escaping(default_return=default_value)
+    @escape(default_return=default_value)
     async def function(a, b, c=5):
         raise ValueError
 
@@ -186,56 +186,56 @@ def test_run_coroutine_function_with_exception_with_default_return():
 
 def test_wrong_argument_to_decorator():
     with pytest.raises(ValueError, match='You are using the decorator for the wrong purpose.'):
-        exception_escaping('kek')
+        escape('kek')
 
 
 def test_context_manager_with_empty_brackets_muted_by_default_exception():
-    with exception_escaping():
+    with escape():
         raise ValueError
 
 
 def test_context_manager_with_empty_brackets_not_muted_by_default_exception():
     for not_muted_exception in (GeneratorExit, KeyboardInterrupt, SystemExit):
         with pytest.raises(not_muted_exception):
-            with exception_escaping():
+            with escape():
                 raise not_muted_exception
 
 
 def test_context_manager_with_exceptions_parameter_not_muted_exception():
     with pytest.raises(ValueError):
-        with exception_escaping(exceptions=(ZeroDivisionError,)):
+        with escape(exceptions=(ZeroDivisionError,)):
             raise ValueError
 
     with pytest.raises(ValueError):
-        with exception_escaping(exceptions=[ZeroDivisionError]):
+        with escape(exceptions=[ZeroDivisionError]):
             raise ValueError
 
 
 def test_context_manager_with_exceptions_parameter_muted_exception():
     for muted_exception in (GeneratorExit, KeyboardInterrupt, SystemExit):
-        with exception_escaping(exceptions=(muted_exception,)):
+        with escape(exceptions=(muted_exception,)):
             raise muted_exception
 
     for muted_exception in (GeneratorExit, KeyboardInterrupt, SystemExit):
-        with exception_escaping(exceptions=[muted_exception]):
+        with escape(exceptions=[muted_exception]):
             raise muted_exception
 
 
 def test_context_manager_without_breackets_muted_exception():
     for muted_exception in (ValueError, KeyError, Exception):
-        with exception_escaping:
+        with escape:
             raise muted_exception
 
 
 def test_context_manager_without_breackets_not_muted_exception():
     for not_muted_exception in (GeneratorExit, KeyboardInterrupt, SystemExit):
         with pytest.raises(not_muted_exception):
-            with exception_escaping:
+            with escape:
                 raise not_muted_exception
 
 
 def test_decorator_without_breackets_saves_name_of_function():
-    @exception_escaping
+    @escape
     def function():
         pass
 
@@ -243,7 +243,7 @@ def test_decorator_without_breackets_saves_name_of_function():
 
 
 def test_decorator_without_breackets_saves_name_of_coroutine_function():
-    @exception_escaping
+    @escape
     async def function():
         pass
 
@@ -252,21 +252,21 @@ def test_decorator_without_breackets_saves_name_of_coroutine_function():
 
 def test_context_manager_with_default_return_value():
     with pytest.raises(SetDefaultReturnValueForDecoratorError, match='You cannot set a default value for the context manager. This is only possible for the decorator.'):
-        with exception_escaping(default_return='lol'):
+        with escape(default_return='lol'):
             ...
 
 def test_set_exceptions_types_with_bad_typed_value():
     with pytest.raises(ValueError, match='The list of exception types can be of the list or tuple type.'):
-        exception_escaping(exceptions='lol')
+        escape(exceptions='lol')
 
 
 def test_set_exceptions_types_with_bad_typed_exceptions_in_list():
     with pytest.raises(ValueError, match='The list of exception types can contain only exception types.'):
-        exception_escaping(exceptions=[ValueError, 'lol'])
+        escape(exceptions=[ValueError, 'lol'])
 
 
 def test_decorator_with_list_of_muted_exceptions():
-    @exception_escaping(exceptions=[ValueError])
+    @escape(exceptions=[ValueError])
     def function():
         raise ValueError
 
@@ -274,7 +274,7 @@ def test_decorator_with_list_of_muted_exceptions():
 
 
 def test_decorator_with_list_of_not_muted_exceptions():
-    @exception_escaping(exceptions=[ValueError])
+    @escape(exceptions=[ValueError])
     def function():
         raise KeyError
 
@@ -283,7 +283,7 @@ def test_decorator_with_list_of_not_muted_exceptions():
 
 
 def test_decorator_with_tuple_of_muted_exceptions():
-    @exception_escaping(exceptions=(ValueError, ))
+    @escape(exceptions=(ValueError, ))
     def function():
         raise ValueError
 
@@ -291,7 +291,7 @@ def test_decorator_with_tuple_of_muted_exceptions():
 
 
 def test_decorator_with_list_of_not_muted_exceptions():
-    @exception_escaping(exceptions=(ValueError,))
+    @escape(exceptions=(ValueError,))
     def function():
         raise KeyError
 
@@ -300,7 +300,7 @@ def test_decorator_with_list_of_not_muted_exceptions():
 
 
 def test_async_decorator_with_list_of_muted_exceptions():
-    @exception_escaping(exceptions=[ValueError])
+    @escape(exceptions=[ValueError])
     async def function():
         raise ValueError
 
@@ -308,7 +308,7 @@ def test_async_decorator_with_list_of_muted_exceptions():
 
 
 def test_async_decorator_with_list_of_not_muted_exceptions():
-    @exception_escaping(exceptions=[ValueError])
+    @escape(exceptions=[ValueError])
     async def function():
         raise KeyError
 
@@ -317,7 +317,7 @@ def test_async_decorator_with_list_of_not_muted_exceptions():
 
 
 def test_async_decorator_with_tuple_of_muted_exceptions():
-    @exception_escaping(exceptions=(ValueError, ))
+    @escape(exceptions=(ValueError, ))
     async def function():
         raise ValueError
 
@@ -325,7 +325,7 @@ def test_async_decorator_with_tuple_of_muted_exceptions():
 
 
 def test_async_decorator_with_list_of_not_muted_exceptions():
-    @exception_escaping(exceptions=(ValueError,))
+    @escape(exceptions=(ValueError,))
     async def function():
         raise KeyError
 
@@ -334,7 +334,7 @@ def test_async_decorator_with_list_of_not_muted_exceptions():
 
 
 def test_default_default_value_is_none():
-    @exception_escaping(exceptions=(ValueError,))
+    @escape(exceptions=(ValueError,))
     def function():
         raise ValueError
 
@@ -342,7 +342,7 @@ def test_default_default_value_is_none():
 
 
 def test_default_default_value_is_none_in_async_case():
-    @exception_escaping(exceptions=(ValueError,))
+    @escape(exceptions=(ValueError,))
     async def function():
         raise ValueError
